+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+<a name="v0.5.0"></a>
+### v0.5.0 - 2019-07-09
+- different styles for inline_code and code_block
+- rgb now longer a macro but a free function
+- two more color functions: ansi and gray
+- clean and complete documentation
+
 <a name="v0.4.0"></a>
 ### v0.4.0 - 2019-07-02
 - support for horizontal rule (line made of dashes)